moved timer - updated models - updated seed file
