+++ src/public/pages/soloRound.html
@@ -55,5 +55,6 @@
 
         <script src="socket.io-client/dist/socket.io.js"></script>
         <script src="soloRound.js"></script>
+        <script src="timer.js"></script>
     </body>
 </html>
